Revert addition of ThrottledPythonSymbolProvider.
