+++ src/client/activation/jedi.ts
@@ -16,7 +16,6 @@ import { PythonReferenceProvider } from '../providers/referenceProvider';
 import { PythonRenameProvider } from '../providers/renameProvider';
 import { PythonSignatureProvider } from '../providers/signatureProvider';
 import { PythonSymbolProvider } from '../providers/symbolProvider';
-import { ThrottledPythonSymbolProvider } from '../providers/throttledSymbolProvider';
 import { IUnitTestManagementService } from '../unittests/types';
 import { WorkspaceSymbols } from '../workspaceSymbols/main';
 import { IExtensionActivator } from './types';
@@ -51,7 +50,7 @@ export class JediExtensionActivator implements IExtensionActivator {
         const serviceContainer = this.serviceManager.get<IServiceContainer>(IServiceContainer);
         context.subscriptions.push(new WorkspaceSymbols(serviceContainer));
 
-        const symbolProvider = new ThrottledPythonSymbolProvider(serviceContainer, jediFactory);
+        const symbolProvider = new PythonSymbolProvider(serviceContainer, jediFactory);
         context.subscriptions.push(languages.registerDocumentSymbolProvider(this.documentSelector, symbolProvider));
 
         const pythonSettings = this.serviceManager.get<IConfigurationService>(IConfigurationService).getSettings();
@@ -61,7 +60,7 @@ export class JediExtensionActivator implements IExtensionActivator {
 
         const testManagementService = this.serviceManager.get<IUnitTestManagementService>(IUnitTestManagementService);
         testManagementService.activate()
-            .then(() => testManagementService.activateCodeLenses(new PythonSymbolProvider(serviceContainer, jediFactory)))
+            .then(() => testManagementService.activateCodeLenses(symbolProvider))
             .catch(ex => this.serviceManager.get<ILogger>(ILogger).logError('Failed to activate Unit Tests', ex));
 
         return true;
